Modify USBtoSerial (USB->RS232) for Adafruit 32u4
To use this, edit the makefile and change the AVRDUDE_PORT location
as appropriate. To program, plug in your breakout over USB, press the white button
and type `make program` to program your breakout.

To get serial output, connect the device you want to get data from to
pin D2 (RXD1 on the Atmel datasheet) and power it on.

Data should come over the specified interface (for Linux this is usually
/dev/ttyACM0).